Kep test files (#290)

+++ pkg/cli/download.go
@@ -18,7 +18,9 @@ import (
 	"fmt"
 	"io"
 	"os"
+	"path/filepath"
 	"regexp"
+	"strings"
 
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
@@ -54,7 +56,7 @@ type nameable interface {
 
 var filenameRegex = regexp.MustCompile("[^A-Za-z0-9]+")
 
-func saveEntitiesLocally[T nameable](saveDir string, entities []T, overrideLocalFiles bool) (int, error) {
+func saveEntitiesLocally[T nameable](saveDir string, entities []T, overrideLocalFiles bool, skipFilesPatterns []string) (int, error) {
 	_, err := os.Stat(saveDir)
 	if err != nil {
 		return 0, err
@@ -68,7 +70,7 @@ func saveEntitiesLocally[T nameable](saveDir string, entities []T, overrideLocal
 		return 0, nil
 	}
 
-	err = purgeDirectory(saveDir)
+	err = purgeDirectory(saveDir, skipFilesPatterns)
 	if err != nil {
 		return 0, fmt.Errorf("could not purge directory %s: %w", saveDir, err)
 	}
@@ -91,16 +93,49 @@ func saveEntitiesLocally[T nameable](saveDir string, entities []T, overrideLocal
 	return saved, nil
 }
 
-// remove all contents of a directory and creates it again
-func purgeDirectory(saveDir string) error {
-	err := os.RemoveAll(saveDir)
-	if err != nil {
-		return fmt.Errorf("error clearing directory %s: %w", saveDir, err)
+// remove all contents of a directory except files matching skipFilesPatterns
+func purgeDirectory(saveDir string, skipFilesPatterns []string) error {
+	// If no skip patterns, remove everything
+	if len(skipFilesPatterns) == 0 {
+		err := os.RemoveAll(saveDir)
+		if err != nil {
+			return fmt.Errorf("error clearing directory %s: %w", saveDir, err)
+		}
+		err = os.MkdirAll(saveDir, 0755)
+		if err != nil {
+			return fmt.Errorf("error creating directory %s: %w", saveDir, err)
+		}
+		return nil
 	}
-	err = os.MkdirAll(saveDir, 0755)
+
+	// Read all files in the directory
+	entries, err := os.ReadDir(saveDir)
 	if err != nil {
-		return fmt.Errorf("error creating directory %s: %w", saveDir, err)
+		return fmt.Errorf("error reading directory %s: %w", saveDir, err)
 	}
+
+	// Check each entry and delete if it doesn't match any skip pattern
+	for _, entry := range entries {
+		entryPath := filepath.Join(saveDir, entry.Name())
+
+		// Check if the entry matches any skip pattern
+		shouldSkip := false
+		for _, pattern := range skipFilesPatterns {
+			if strings.Contains(entry.Name(), pattern) {
+				shouldSkip = true
+				break
+			}
+		}
+
+		// Delete if it doesn't match any skip pattern
+		if !shouldSkip {
+			err := os.RemoveAll(entryPath)
+			if err != nil {
+				return fmt.Errorf("error removing %s: %w", entryPath, err)
+			}
+		}
+	}
+
 	return nil
 }
